+++ plugins/woocommerce/tests/e2e-pw/tests/merchant/products/block-editor/disable-block-product-editor.spec.js
@@ -1,6 +1,8 @@
 const { test, expect } = require( '@playwright/test' );
 const {
 	clickAddNewMenuItem,
+	expectBlockProductEditor,
+	expectOldProductEditor,
 	isBlockProductEditorEnabled,
 	toggleBlockProductEditor,
 } = require( '../../../../utils/simple-products' );
@@ -27,18 +29,6 @@ async function dismissFeedbackModalIfShown( page ) {
 			.getByRole( 'button', { name: 'Skip' } )
 			.click( { timeout: 5000 } );
 	} catch ( error ) {}
-}
-
-async function expectOldProductEditor( page ) {
-	await expect(
-		page.locator( '#woocommerce-product-data h2' )
-	).toContainText( 'Product data' );
-}
-
-async function expectBlockProductEditor( page ) {
-	await expect(
-		page.locator( '.woocommerce-product-header__inner h1' )
-	).toContainText( 'Add new product' );
 }
 
 test.describe( 'Disable block product editor', () => {
