+++ plugins/woocommerce/tests/e2e-pw/utils/simple-products.js
@@ -1,3 +1,5 @@
+const { expect } = require( '@playwright/test' );
+
 const SETTINGS_URL =
 	'wp-admin/admin.php?page=wc-settings&tab=advanced&section=features';
 
@@ -54,7 +56,31 @@ async function clickAddNewMenuItem( page ) {
 		.click();
 }
 
+/**
+ * This function checks if the old product editor is visible.
+ *
+ * @param {Page} page - Represents a single tab in a browser, controlled by puppeteer.
+ */
+async function expectOldProductEditor( page ) {
+	await expect(
+		page.locator( '#woocommerce-product-data h2' )
+	).toContainText( 'Product data' );
+}
+
+/**
+ * This function checks if the block product editor is visible.
+ *
+ * @param {Page} page
+ */
+async function expectBlockProductEditor( page ) {
+	await expect(
+		page.locator( '.woocommerce-product-header__inner h1' )
+	).toContainText( 'Add new product' );
+}
+
 module.exports = {
+	expectBlockProductEditor,
+	expectOldProductEditor,
 	clickAddNewMenuItem,
 	isBlockProductEditorEnabled,
 	toggleBlockProductEditor,
